pyproject.toml
[tool.poetry]
name = "dynamic-llamafile-docker"
version = "1.0.0"
description = "A dynamic strategy to create a Llamafile Docker image through some predefined gguf models"
authors = ["Thiago Costa <thiago.gcosta@alumni.usp.br>"]
maintainers = ["Thiago Costa <thiago.gcosta@alumni.usp.br>"]
readme = "README.md"
license = "MIT"
keywords = ["llamafile", "docker image", "dynamic llamafile", "dynamic llamafile docker"]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9"
wget = "3.2"
pydantic = "2.9.2"
pyyaml = "6.0"
openai = "1.47.0"

[tool.poetry.dev-dependencies]
pre-commit = "3.8.0"
pytest = "8.3.3"
pytest-cov = "5.0.0"

#REF: https://pre-commit.com/
#REF: https://stefaniemolin.com/articles/devx/pre-commit/setup-guide/
#REF: https://medium.com/@pijpijani/smarter-commits-with-git-pre-commit-hooks-9c7bb3f699bf
[tool.ruff]
line-length = 88
select = [
    "C",  # mccabe rules
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
ignore = [
    "C901",  # max-complexity-10
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.numpydoc_validation]
checks = [
    "all",  # report on all checks
    "ES01", # but don't require an extended summary
    "EX01", # or examples
    "SA01", # or a see also section
    "SS06", # and don't require the summary to fit on one line
]
exclude = [  # don't report on checks for these
    '\\.__init__$',
    '\\.__repr__$',
    '\\.__str__$',
]
override_SS05 = [  # allow docstrings to start with these words
    '^Process ',
    '^Assess ',
    '^Access ',
]
